refactor: BE 아키텍처 코드 중복 제거 및 서비스 분리 (#170, #176, #179, #199)
- utils/date_utils.py 신규: expenses.py, income.py 공통 날짜 유틸 추출 (#199)
- schemas/base.py 신규: coerce_date_to_datetime validator mixin (#179)
  - ExpenseBase, IncomeBase 동일 validator 중복 제거
- services/budget_service.py 신규: 예산 알림/카테고리 개요 비즈니스 로직 분리 (#176)
  - get_budget_alerts, get_category_overview → budget_service로 이동
  - budget.py API는 얇은 래퍼로 단순화
