+++ backend/app/schemas/expense.py
@@ -3,29 +3,17 @@
 from datetime import datetime
 from decimal import Decimal
 from enum import Enum
-from typing import Any
 
-from pydantic import BaseModel, ConfigDict, Field, field_validator
+from pydantic import BaseModel, ConfigDict, Field
+
+from app.schemas.base import DateTimeCoerceMixin
 
 
-class ExpenseBase(BaseModel):
+class ExpenseBase(DateTimeCoerceMixin):
     amount: Decimal = Field(..., gt=0, description="지출 금액 (0보다 커야 함)")
     description: str
     category_id: int | None = None
     date: datetime
-
-    @field_validator("date", mode="before")
-    @classmethod
-    def coerce_date_to_datetime(cls, v: Any) -> Any:
-        """YYYY-MM-DD 형식(시간 없는 날짜)을 datetime으로 변환
-
-        Pydantic v2는 기본적으로 날짜만 있는 문자열을 datetime으로 파싱하지 못함.
-        프론트엔드 date input (YYYY-MM-DD) 및 LLM이 반환하는 날짜 형식과 호환되도록
-        T00:00:00을 자동으로 추가한다.
-        """
-        if isinstance(v, str) and len(v) == 10 and "T" not in v and " " not in v:
-            return f"{v}T00:00:00"
-        return v
 
 
 class ExpenseCreate(ExpenseBase):
